requirements.txt
orbis>=2.0
orbis_plugin_aggregation_dbpedia_entity_types>=1.0
pyspotlight==0.7.2
